refactor(app): split main.rs into focused modules
Extract 697-line main.rs into four focused modules:
- scene.rs: terrain generation, particle spawning, material palette
- app.rs: App struct, ApplicationHandler impl, input handling
- headless.rs: headless execution mode
- main.rs: entry point, arg parsing, mod declarations

Closes #186
